Clean up regex logic
regEndsWithFlags.test always returns true because it uses * instead
of + after the list of flags. We could fix that, but it turns out
the test is unnecessary anyway -- the logic for the branch works
whether the flags are empty or not. And the logic in the 'else' branch
was actually broken.